+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/storage/annotation/ValueColumnMetadata.java
@@ -24,7 +24,6 @@ import java.util.Optional;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.apache.skywalking.oap.server.core.query.enumeration.Scope;
-import org.apache.skywalking.oap.server.core.query.sql.Function;
 
 /**
  * ValueColumnMetadata holds the metadata for column values of metrics. The metadata of ValueColumn is declared through
@@ -42,10 +41,9 @@ public enum ValueColumnMetadata {
     public void putIfAbsent(String modelName,
                             String valueCName,
                             Column.ValueDataType dataType,
-                            Function function,
                             int defaultValue,
                             int scopeId) {
-        mapping.putIfAbsent(modelName, new ValueColumn(valueCName, dataType, function, defaultValue, scopeId));
+        mapping.putIfAbsent(modelName, new ValueColumn(valueCName, dataType, defaultValue, scopeId));
     }
 
     public void overrideColumnName(String oldName, String newName) {
@@ -58,13 +56,6 @@ public enum ValueColumnMetadata {
     public String getValueCName(String metricsName) {
         final String valueCName = findColumn(metricsName).valueCName;
         return columnNameOverrideRule.getOrDefault(valueCName, valueCName);
-    }
-
-    /**
-     * Fetch the function for the value column of the given metrics name.
-     */
-    public Function getValueFunction(String metricsName) {
-        return findColumn(metricsName).function;
     }
 
     public int getDefaultValue(String metricsName) {
@@ -102,7 +93,6 @@ public enum ValueColumnMetadata {
     public static class ValueColumn {
         private final String valueCName;
         private final Column.ValueDataType dataType;
-        private final Function function;
         private final int defaultValue;
         private final int scopeId;
     }